added a simple textbox in the frontend
found a cool textbox for the frontend so i added it in components folder for frontend team to use.

added a components folder where components will be (like textbox.jsx)

created file app.js in frontend to get data with functions from backend.

contact me for full explanantion of updates if req