Added debouncing, cleaned up code, separated some css, added a custom hook
Recently heard of debouncing so implemented it in the search bar to help reduce api calls on every keystroke

separated search bar from the nav menu in order to make it reusable for searching users or for searching stories - the search bar is it's own component now.

created get users function in app.js to get all users and search for them.

separated annoying css where cluttering file code (search bar)

got username and added it to top of swipe bar on mobile version to know who is signed in.